requirements.txt
Django==2.1.2
django-cors-middleware==1.3.1
django-extensions==2.1.2
django-inlinecss==0.1.2
django-rest-swagger==2.2.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
psycopg2-binary==2.7.5
gunicorn==19.9.0
psycopg2==2.7.5
PyJWT==1.6.4
pytz==2018.5
six==1.11.0
whitenoise==4.1
flake8==3.5.0
cloudinary==1.13.0
Pillow==5.3.0
Faker==0.9.1
social-auth-app-django==2.1.0
python-social-auth==0.3.6
django-filter==2.0.0
django-notifications-hq==1.5.0
